refactor(dir): rename filesystem provider module
